+++ unit_test/namespace_test.go
@@ -7,7 +7,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package modusdb_test
+package unit_test
 
 import (
 	"context"
